requirements.txt
PyAutoGUI==0.9.54
PyYAML==6.0.1
